# backend/openapi.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/food-donation":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFoodDonationDto"}}}},"responses":{"201":{"description":"Collecte créé."}},"summary":"Créer une collecte","tags":["FoodDonation"]},"get":{"operationId":"findAll","parameters":[],"responses":{"200":{"description":"Past and upcoming food donations","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FoodDonationListResponse"}}}}},"summary":"Récupérer toutes les collectes","tags":["FoodDonation"]}},"/food-donation/{id}":{"get":{"operationId":"findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["FoodDonation"]},"patch":{"operationId":"update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateFoodDonationDto"}}}},"responses":{"200":{"description":""}},"tags":["FoodDonation"]},"delete":{"operationId":"remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["FoodDonation"]}},"/booking":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookingDto"}}}},"responses":{"201":{"description":"Réservation crée"}},"summary":"Créer une réservation","tags":["Booking"]},"get":{"operationId":"findAll","parameters":[],"responses":{"200":{"description":"Une liste de réservation","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Booking"}}}}}},"tags":["Booking"]}},"/booking/{id}":{"get":{"operationId":"findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Un booking","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Booking"}}}}},"summary":"Récupérer une réservation","tags":["Booking"]}},"/association":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAssociationDto"}}}},"responses":{"201":{"description":"Association créé."}},"summary":"Créer une association","tags":["Association"]},"get":{"operationId":"findAll","parameters":[],"responses":{"200":{"description":"Les associations","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Association"}}}}}},"summary":"Récupérer toutes les assos","tags":["Association"]}},"/association/{id}":{"get":{"operationId":"findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Une association","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Association"}}}}},"summary":"Récupérer une asso","tags":["Association"]}}},"info":{"title":"Save meal API","description":"","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateFoodDonationDto":{"type":"object","properties":{"title":{"type":"string"},"foodType":{"type":"string"},"estimatedPortions":{"type":"number"},"description":{"type":"string"},"pickupPlace":{"type":"string"},"address":{"type":"string"},"pickupInstructions":{"type":"string"},"availableFrom":{"type":"string"},"availableTo":{"type":"string"},"contactName":{"type":"string"},"contactPhone":{"type":"string"},"contactEmail":{"type":"string"},"additionalNotes":{"type":"string","nullable":true},"createdAt":{"type":"string","nullable":true},"updatedAt":{"type":"string","nullable":true}},"required":["title","foodType","estimatedPortions","description","pickupPlace","address","pickupInstructions","availableFrom","availableTo","contactName","contactPhone","contactEmail"]},"FoodDonationListResponse":{"type":"object","properties":{"past":{"type":"array","items":{"type":"string"}},"upcoming":{"type":"array","items":{"type":"string"}}},"required":["past","upcoming"]},"UpdateFoodDonationDto":{"type":"object","properties":{}},"CreateBookingDto":{"type":"object","properties":{}},"Booking":{"type":"object","properties":{"id":{"type":"number"},"associationId":{"type":"number"},"foodDonationId":{"type":"number"},"createdAt":{"format":"date-time","type":"string"}},"required":["id","associationId","foodDonationId","createdAt"]},"CreateAssociationDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"}},"required":["id","name"]},"Association":{"type":"object","properties":{}}}}}