# web/api-client/src/models/CreateAssociationDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Save meal API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateAssociationDto
 */
export interface CreateAssociationDto {
    /**
     * 
     * @type {number}
     * @memberof CreateAssociationDto
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof CreateAssociationDto
     */
    name: string;
}

/**
 * Check if a given object implements the CreateAssociationDto interface.
 */
export function instanceOfCreateAssociationDto(value: object): value is CreateAssociationDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function CreateAssociationDtoFromJSON(json: any): CreateAssociationDto {
    return CreateAssociationDtoFromJSONTyped(json, false);
}

export function CreateAssociationDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateAssociationDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
    };
}

export function CreateAssociationDtoToJSON(json: any): CreateAssociationDto {
    return CreateAssociationDtoToJSONTyped(json, false);
}

export function CreateAssociationDtoToJSONTyped(value?: CreateAssociationDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
    };
}