# web/api-client/src/models/FoodDonationListResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Save meal API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FoodDonationListResponse
 */
export interface FoodDonationListResponse {
    /**
     * 
     * @type {Array<string>}
     * @memberof FoodDonationListResponse
     */
    past: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof FoodDonationListResponse
     */
    upcoming: Array<string>;
}

/**
 * Check if a given object implements the FoodDonationListResponse interface.
 */
export function instanceOfFoodDonationListResponse(value: object): value is FoodDonationListResponse {
    if (!('past' in value) || value['past'] === undefined) return false;
    if (!('upcoming' in value) || value['upcoming'] === undefined) return false;
    return true;
}

export function FoodDonationListResponseFromJSON(json: any): FoodDonationListResponse {
    return FoodDonationListResponseFromJSONTyped(json, false);
}

export function FoodDonationListResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): FoodDonationListResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'past': json['past'],
        'upcoming': json['upcoming'],
    };
}

export function FoodDonationListResponseToJSON(json: any): FoodDonationListResponse {
    return FoodDonationListResponseToJSONTyped(json, false);
}

export function FoodDonationListResponseToJSONTyped(value?: FoodDonationListResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'past': value['past'],
        'upcoming': value['upcoming'],
    };
}